# DSell_API_OpenAPI.json
{
    "openapi":  "3.0.3",
    "info":  {
                 "title":  "DSell 直销管理系统 API",
                 "version":  "1.0.0",
                 "description":  "Base URL: http://localhost:8080\n认证方式: Header 中携带 token (登录接口返回)"
             },
    "servers":  [
                    {
                        "url":  "http://localhost:8080"
                    }
                ],
    "security":  [
                     {
                         "token":  [

                                   ]
                     }
                 ],
    "components":  {
                       "securitySchemes":  {
                                               "token":  {
                                                             "type":  "apiKey",
                                                             "in":  "header",
                                                             "name":  "token"
                                                         }
                                           },
...
                   },
...
    "paths":  {
...
                  "/api/system/user/page":  {
                                                "get":  {
                                                            "tags":  [
                                                                         "System"
                                                                     ],
                                                            "summary":  "分页查询用户",
                                                            "parameters":  [
                                                                               {
                                                                                   "name":  "page",
                                                                                   "in":  "query",
                                                                                   "schema":  {
                                                                                                  "type":  "integer",
                                                                                                  "default":  1
                                                                                              }
                                                                               },
                                                                               {
                                                                                   "name":  "size",
                                                                                   "in":  "query",
                                                                                   "schema":  {
                                                                                                  "type":  "integer",
                                                                                                  "default":  10
                                                                                              }
                                                                               }
                                                                           ]
                                                        }
                                            },
                  "/api/system/user/{id}":  {
                                                "get":  {
                                                            "tags":  [
                                                                         "System"
                                                                     ],
                                                            "summary":  "获取用户详情",
                                                            "parameters":  [
                                                                               {
                                                                                   "name":  "id",
                                                                                   "in":  "path",
                                                                                   "required":  true,
                                                                                   "schema":  {
                                                                                                  "type":  "integer"
                                                                                              }
                                                                               }
                                                                           ]
                                                        },
                                                "delete":  {
                                                               "tags":  [
                                                                            "System"
                                                                        ],
                                                               "summary":  "删除用户",
                                                               "parameters":  [
                                                                                  {
                                                                                      "name":  "id",
                                                                                      "in":  "path",
                                                                                      "required":  true,
                                                                                      "schema":  {
                                                                                                     "type":  "integer"
                                                                                                 }
                                                                                  }
                                                                              ]
                                                           }
                                            },
...
                  "/api/system/user/{id}/status":  {
                                                       "put":  {
                                                                   "tags":  [
                                                                                "System"
                                                                            ],
                                                                   "summary":  "修改用户状态",
                                                                   "parameters":  [
                                                                                      {
                                                                                          "name":  "id",
                                                                                          "in":  "path",
                                                                                          "required":  true,
                                                                                          "schema":  {
                                                                                                         "type":  "integer"
                                                                                                     }
                                                                                      },
                                                                                      {
                                                                                          "name":  "status",
                                                                                          "in":  "query",
                                                                                          "required":  true,
                                                                                          "schema":  {
                                                                                                         "type":  "integer",
                                                                                                         "enum":  [
                                                                                                                      0,
                                                                                                                      1
                                                                                                                  ]
                                                                                                     }
                                                                                      }
                                                                                  ]
                                                               }
                                                   },
...
              }
}